Tests moved into separate files
